+++ src/meta/area/index.ts
@@ -7,3 +7,4 @@ export { Global } from './global'
 export type { Region } from './region'
 export { RegionCode } from './regionCode'
 export type { RegionGroup } from './regionGroup'
+export { AssessmentStatus } from './status'
