+++ src/meta/user/userRoles.ts
@@ -1,8 +1,7 @@
 import i18n from 'i18next'
 import { Objects } from 'utils/objects'
 
-import { Areas } from 'meta/area'
-import { AssessmentStatus } from 'meta/area/country'
+import { Areas, AssessmentStatus } from 'meta/area'
 import { Assessment } from 'meta/assessment'
 import { User } from 'meta/user/user'
